requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
Pillow==8.3.1
markdown2==2.4.0